thermal: qcom: simple: Only consider LITTLE cluster temperature
Utilizing big and LITTLE clusters turned out to be bad because the overall temperature were bigger than necessary, leading to constant throttling. To fix that, we simply take in consideration only the LITTLE cluster. In this CPU (6+2), there are more LITTLE cores than big cores so this makes more sense as they are the majority of the CPUs.

Signed-off-by: Yahya Wessam <yahyawessam2002@gmail.com>